Move PHPUnit related code to PHPUnitSkeleton namespace
